Move DataSourcePicker implementation to core
